api/requirements.txt
asgiref
Django 
django-cors-headers 
djangorestframework
gunicorn
mysqlclient
pytz
colorama
black
faker